[Refactoring] functions.py -> utils.py 로 파일명 변경
